device-tree: Consolidate device-tree recipe and append
With the introduction of devicetree.bbclass in oe-core/meta, remove the
implementation of device tree compilation from device-tree.bb keeping
the meta-xilinx specific information and licensing. Also consolidate the
device-tree.bb and device-tree.bbappend.

The only differences between the existing device-tree.bb implementation
and the devicetree.bbclass implementation is the device trees are
deployed in a "devicetree/" subdirectory within the deployed images
directory.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>