arm-trusted-firmware.inc: Add support for atf to use 2nd UART
Add ZYNQMP_CONSOLE flag to EXTRA_OEMAKE to add support for the 2nd UART.
If ZYNQMP_CONSOLE is set to cadence or cadence0, it will use UART 0 and
if set to cadence1, it will use UART 1. The default value is set to
cadence.

Change do_configure to perform make clean. Without this change,you must
manually clean sstate before building with the new value of
ZYNQMP_CONSOLE, which is not the desired behavior.

Signed-off-by: Jaewon Lee <jaewon.lee@xilinx.com>
Signed-off-by: Bhargava Sreekantappa Gayathri <bhargava.sreekantappa-gayathri@xilinx.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>