recipes-multimedia/vcu/*: Set package arch to SOC_FAMILY
Set package arch of VCU to SOC_FAMILY. Even though VCU hwcodec is
available only in ev devices, we are enabling the package for all MPSoC
devices.

gstreamer is the application used for video pipelining. When the VCU HW
is not found, the behaviour will default to using software based
gstreamer (in case of cg and eg devices).

By making package arch common for all zynqmp devices, we can use the
same sstate-cache and RPM feeds for gstreamer components.

Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>