qemu multiarch: Replace shell based wrapper with python one
The qemu-system-aarch64-multiarch wrapper parses the arguments
passed to it, and launches two separate instances of qemu
with the correct arguments for PMU and APU respectively.

Using a python based wrapper gives us more control over
how the calls to the qemu instances are made and how the
arguments are passed.

This patch replaces the old wrapper with a python based one
providing the same functionality as the old one but also
enables testimage to be run on qemu instances, switching the
tcp serial ports passed to qemu to be able to read the output
from qemurunner correctly

Signed-off-by: Alejandro Enedino Hernandez Samaniego <alejandr@xilinx.com>