u-boot-zynq-uenv: Create dependency from images to uEnv file
Images include a uEnv.txt file that is read from U-Boot,
these parameters are read after the default environment
is loaded on U-Boot and just before booting the OS.

Changing anything from the environment through the build
system created a new uEnv file, but this didnt cause the
build system to create a new image (that included the
new file), so changes did not reflect on the image until
a new build was executed.

This patch creates a dependency to u-boot-zynq-uenv:do_deploy
which will cause the image to be recreated when something
from uEnv has changed.

Signed-off-by: Alejandro Enedino Hernandez Samaniego <alejandr@xilinx.com>