meta-xilinx-contrib/u-boot: Remove minized appends
The minized board is supported in the v2018.07 of u-boot which is the
default provided by oe-core/meta. This includes all the configuration
provided by the patch.

Note however that neither upstream u-boot nor the patch in
meta-xilinx-contrib provided ps7 init files for this board which makes
the HAS_PLATFORM_INIT incorrect, as it will result in a boot.bin being
built which does not include ps7 init configuration and thus will not
boot correctly. As such this change also removes the boot.bin setup and
use in the machine configuration to prevent the dependency on platform
init files which are not available.

Signed-off-by: Nathan Rossi <nathan@nathanrossi.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>