pmu-firmware: Port pmu-firmware recipe
This patch ports the pmu-firmware recipe from meta-xilinx-bsp to be used
with the standalone/baremetal toolchain and also upgrades it to the
latest release at this point.

The recipe was trimmed down, and a few changes had to be made to make it
compatible with the baremetal layer, DEPENDS, pass include dir, license
and such.

Signed-off-by: Alejandro Enedino Hernandez Samaniego <alejandr@xilinx.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>