Enforced provider for boot-bin
This provider isn't being automatically picked up and the
boot image dependency on a boot.bin file is also missing.
Given this branch is intended to be used in sync with the
one from meta-xilinx-tools, and related missing support
for a bootable U-Boot SPL noted repeatedly in meta-xilinx's
mailing list, we're setting meta-xilinx-tools provider as
the default.  We've tested the zcu111 and zcu111 with this
in concert with changes to meta-xilinx-tools to provide
the file in the right name to not be picked up by the uEnv
tooling.

Signed-off-by: Thomas Goodwin <btgoodwin@geontech.com>