libmali-xlnx: Use update-alternatives to switch between GL backends
libmali provides GL backends based on x11, fbdev and wayland,
we should be able to switch between them at runtime since it is
the same ABI, it should only be a matter of loading the correct
shared library by the interpreter (dynamic linker).

Use the update-alternatives class, to provide a way for the user
to choose the desired backend at runtime, do this by setting
priorities, the package with the highest priority will be chosen
as default at build time, but it can easily be changed at runtime
afterwards.
This change implies that the libmali package will install all
backends regardless of which one was chosen, but it will only use
one as default.

Use the x11 backend by default at build time; given that it is
the same ABI, applications which depend on libmali, can build
regardless of the chosen backend at build time.

Update-alternatives uses a set of commands on the postinst
scripts when creating the root filesystem, which basically create
the soft link between the chosen alternative and the binary/library.
This usually works seamlessly (for binaries), but it does not in the
case of libraries, because ldconfig is run at the end of the
do_rootfs task, and it removes the link that was just created,
it is important to note that this is simply normal ldconfig behavior
and its not something we can fix, so we defer execution of
update-alternatives until the first boot, hence avoiding the link
removal by ldconfig.

Switching backends at build time will also help to avoid longer
build times, since it will only invalidate the do_package task,
rebuilding an image after switching a backend (at build time)
should only execute the do_package task along with the do_rootfs
task.

Signed-off-by: Alejandro Enedino Hernandez Samaniego <alejandr@xilinx.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>