opencl-clhpp: Allow empty packages to be built
The opencl-clhpp package simply contains headers, hence they are all
contained within the opencl-clhpp-dev package.

This causes the opencl-clhpp to be empty and the default behavior is to
not create empty packages, although due to rpm packaging behavior a
dependency exists between the opencl-clhpp-dev package and the
opencl-clhpp, which causes an error when creating the rootfs.

Use ALLOW_EMPTY_${PN}= "1" to allow the empty package to be created,
that way dependencies are fullfilled when creating the rootfs.

Signed-off-by: Alejandro Enedino Hernandez Samaniego <alejandr@xilinx.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>