run-postinsts: Pass the output of the scripts run to kmsg
The run-postinsts scripts are meant to run during the first boot of the device.

At boot time, though, some resources are not available(e.g. STDOUT), hence
why applications that run at this point need to be correctly daemonized,
if, for some reason a script that runs is not able to start as a daemon,
proper behavior cannot be guaranteed.

For some unknown reason to me at this point, the console behaves differenly
on our devices depending on the chosen boot mode, for example, when using
sdboot, scripts that echo something during boot show the messages properly
on the boot log, but when using jtag as boot mode, this does not happen and
the messages are not shown.

One of such cases happens with update-alternatives, which expects STDOUT
to be present, and if it fails to get a possible output file descriptor it
exits ungracefully, causing the run-postinsts script to exit itself.

Pass the the kmsg character device as the output for the scripts run by
the run-postinsts script, not only to be able to print out the output of
such script to the boot log, but also to pass a valid file descriptor
to the scripts so they behave properly.

Signed-off-by: Alejandro Enedino Hernandez Samaniego <alejandr@xilinx.com>