kernel-module-hdmi_git.bb: New Yocto recipe for Xilinx HDMI drivers
This recipe will pull in the Xilinx HDMI drivers and associated
components, for soft IP, to be built as an out-of-tree modules.

Following drivers will be pulled-in

hdmi-rx (v4l2):
The HDMI 1.4/2.0 Receiver Subsystem is a feature-rich soft IP
incorporating all the necessary logic to properly interface with PHY
layers and provide HDMI decoding functionality. The subsystem is a
hierarchical IP that bundles a collection of HDMI RX-related IP
sub-cores and outputs them as a single IP.  The subsystem receives the
captured TMDS data from the video PHY layer. It then extracts the video
and audio streams from the HDMI stream and converts it to AXI video and
audio streams. It includes support for HDCP 1.4 and 2.2 protocols

hdmi-tx (drm):
HDMI Tx subsystem is a feature-rich soft IP incorporating all the
necessary logic to properly interface with xilinx PHY layers and provide
HDMI encoding functionality. The subsystem is a hierarchical IP that
bundles a collection of HDMI TX-related IP sub-cores and outputs them as
a single IP. The subsystem takes incoming video and audio streams and
transfers them to an HDMI stream. The stream is then forwarded to the
video PHY layer. It includes support for HDCP 1.4 and 2.2 protocols

video-phy (phy):
Xilinx Video Phy implements the physical layer for enabling the
plug-and-play connectivity with HDMI MAC transmit and receive
subsystems.  This driver is also used as the repository for common files
used by hdmi rx and tx drivers including hdcp1.4 and hdcp2.2 software
stack

si5324 (ccf):
Silicon labs clock generator driver specifically targeting the xilinx
hdmi use-case. i.e. implemented only functionality required for xilinx
hdmi soft IP at this time.

dp159 (ccf):
SNxDP159 device is a dual mode display port to TMDS retimer supporting
DVI 1.0 and HDMI 1.4b and 2.0 output signals. When working as a retimer,
the embedded clock data recovery cleans up the input high frequency and
random jitter from video source. It also supports TMDS output amplitude
adjust and output slew rate control.

Signed-off-by: Rohit Consul <rohit.consul@xilinx.com>
Signed-off-by: Shikhar Mishra <shikhar.mishra@xilinx.com>
Signed-off-by: Manjukumar Matha <manjukumar.harthikote-matha@xilinx.com>
Signed-off-by: Sai Hari Chandana Kalluri <chandana.kalluri@xilinx.com>