zynqmp-fsbl: Fix race condition on copy_psu_init
Theres a race condition that happens when the copy_psu_init task
is executed, and claims that the src directory does not exist, when
this task is trying to access it.

There is no need to create an additional task to copy the psu_init code,
it can simply be prepended to do_configure which achieves the same result
without having to keep track of other tasks dependencies that might
cause an error.

Copy the psu_init files on do_configure itself avoiding the race condition.

Signed-off-by: Alejandro Enedino Hernandez Samaniego <alejandr@xilinx.com>