meta-xilinx-standalone: Using S instead of WORKDIR
External source class manipulates S to point to external source directories,
so use this variable instead of WORKDIR. Point S to the top directory as we need to
access scripts directory in top directory. And instead, using OECMAKE_SOURCEPATH to
point to deeper driver specific directory with necessary cmake files.

Signed-off-by: Jaewon Lee <jaewon.lee@xilinx.com>
Signed-off-by: Appana Durga Kedareswara rao <appana.durga.rao@xilinx.com>