device-tree: Fix install directory
There was a typo in the patch that installed the psu_init code
which initially tried to install files on the /var directory.

Since this is a baremetal component, there is basically no
convention on where files should be installed, for now
to comply a bit more with bitbake, install these files on
includedir to fix the install-vs-shipped issue.

Signed-off-by: Alejandro Enedino Hernandez Samaniego <alejandr@xilinx.com>