gcc-cross-canadian_%.bbappend:temporary hack to build gcc cross canadian
This is a TEMPORARY HACK to build gcc-cross-canadian toolchain for the
xilinx-standalone distro.

xilinx-standalone.conf will set TCLIB as newlib which will pull in
tclibc-newlib.inc which sets TARGET_OS to eabi for arm and elf for
everything else. In gcc-cross-canadian.inc (upstream) there is
a _remove('--with-sysroot=/not/exist') for TARGET_OS overrides(elf and
eabi) which fixes baremetal toolchain builds by preventing the compiler
from looking in nonexistent sysroot directories
e.g. https://github.com/riscv/meta-riscv/issues/117

But that flag is needed when building for newlib as that is what will
specify the path in the arch specific sysroots to look for the header
files that were generated.

Without setting --with-sysroot the following error would
occur:
    fatal error: stdio.h: No such file or directory
    1 | #include <stdio.h>

Signed-off-by: Jaewon Lee <jaewon.lee@xilinx.com>
Signed-off-by: Alejandro Enedino Hernandez Samaniego <alejandr@xilinx.com>