meta-xilinx-standalone: recipes-applications: hello-world: Remove dependency on esw_examples class
Hello world recipe doesn't have any dependency on the
do_generate_eglist task, It has dependency only on the linker generation.
So the esw repo linker generation is moved to the liner_gen.py
file, which needs to be called before do_configure.

This patch updates the recipe for the same.

Signed-off-by: Appana Durga Kedareswara rao <appana.durga.rao@xilinx.com>