SOC_FAMILY: Change overrides
The ESW repo requires to pass the application that is being built
along with the proessor architecture (family) where it will be
executed to workaround some of the driver/libraries functionality.

For us to be able to override variables depending on the application
we need to create separate machines depending on the family that will
be used, e.g. microblaze-pmu and microblaze-plm, or family where it
could be zynqmp or versal, these are passed by the XLNX_CMAKE_PROCESSOR
and XLNX_CMAKE_MACHINE variables respectively, these separate machines
will be built using separate multiconfigs, and while this is an
expensive solution, since it costs: parsing and building time along with
storage use for each of the multiconfigs, this is the correct solution
due to their incompatibility on their dependencies.

As part of these changes, it is also required to change the existing
overrides for the MACHINES that already existed for them to keep the
existing functionality on the standalone toolchain.

Signed-off-by: Alejandro Enedino Hernandez Samaniego <alejandr@xilinx.com>