Fix arm cortex r/m profiles
As part of the merge of the code bases, changes were introduced that
changed the way the cortexr5 was defined.  To sync this between the
baremetal compilation and the baremetal toolchain compilation we
create a new armrm tune file, define a new tune feature of 'armrm',
define 'armrm' as a new machine override based on the feature 'armrm',
and move the cortexr5 tuning to a common file.

Signed-off-by: Mark Hatle <mark.hatle@xilinx.com>