xilinx-standalone.conf: Adding qemu to TOOLCHAIN_HOST_TASK
When running gcc built for xilinx-standalone using mingw sdk machine, it
would throw an error saying the winpthread dll was missing. this is
because in tclibc-newlib, TOOLCHAIN_HOST_TASK was being overwritten and
omitting nativesdk-packagegroup-sdk-host which pulled in the appropriate
package nativesdk-mingw-w64-winpthreads. Manually adding the following
to TOOLCHAIN_HOST_TASK to bypass this issue:
nativesdk-qemu-xilinx

Signed-off-by: Jaewon Lee <jaewon.lee@xilinx.com>
Signed-off-by: Alejandro Enedino Hernandez Samaniegoi <alejandr@xilinx.com>