newlib: Adjust configuration for standalone to allow BSP library
Both libgloss and newlib require the same arguents to be passed in.  So
keep the two files in sync with each other.

--disable-newlib-supplied-syscalls

This configuration argument tells newlib to NOT provide builtin
syscalls, which allows the BSP library to provide them.

Note linking to a BSP library will require a custom GCC spec file when
linking.

Signed-off-by: Mark Hatle <mark.hatle@xilinx.com>