system-zcu102: Create heterogeneous machine configuration for ZCU102 evaluation board.
Using the MACHINE=system-zcu102 should allow users to build the required artifacts
for the several architectures inside a ZCU102 board.

This machine along with the declared multiconfigs set up the wiring to do so,
building Linux, PMU firmware and FSBL on a single bitbake invokation:

$ bitbake core-image-minimal

Three different TMP directories will be used, these depend both on the architecture
for which packages are being generated and the C library that was used for that
configuration.

This should be used as an example that can easily be adapted to other boards.

Signed-off-by: Alejandro Enedino Hernandez Samaniego <alejandr@xilinx.com>