xlnx-compatible-os.bbclass: Class to allow recipes to list OS compatibility
Following the example of base.bbclass in OE-Core, a new COMPATIBLE_OS
variable is defined.  Similar to COMPATIBLE_MACHINE, it is a regex that
can be used to declare TARGET_OS string compatibility on a per-recipe
basis.  Either the distro configuration or the recipes themselves will
need to inherit this class.

By default the class assumes we're building a Linux based system, and
thus any recipe who has not declared compatibility is compatible with
Linux.  The default compatible field was copied from bitbake.conf to
ensure there are no surprised.

Additionally, enable the usage of this class throughout
meta-xilinx-standalone layer.

Signed-off-by: Mark Hatle <mark.hatle@xilinx.com>