microblaze_dtb.py: Convert a dtb to one or more microblaze TUNE_FEATURES
Because the microblaze is adjustable on the FPGA, we need a way for
the system to adjust the TUNE_FEATURES for a given build.  The
device tree can specify the capabilities of the microblaze
implementations defined in a specific design.  If present, these can
then be processed to provide a set of TUNE_FEATURES for each declared
microblaze CPU.

The output of the tool is formated in a way that it can be directly
included in the user's local.conf, or placed in a file that is included
by a machine or other file.  The user will only have to declare the
DEFAULTTUNE value that matches one of the generated tunes.

General rules that are implemented:

  TUNE_FEATURE         CONDITION
  ------------         -------------------------------
  microblaze           (all microblaze cpus)
  bigendian            xlnx,endianness is not 1
  64-bit               xlnx,data-size is 64
  barrel-shift         xlnx,use-barrel is 1
  pattern-compare      xlnx,use-pcmp-instr is 1
  reorder              xlnx,use-reorder-instr is not 0
  frequency-optimized  xlnx,area-optimized is 2
  multiply-low         xlnx,use-hw-mul is 1
  multiple-high        xlnx,use-hw-mul is 2
  divide-hard          xlnx,use-div is 1
  fpu-soft             xlnx,use-fpu is not 1 or 2
  fpu-hard             xlnx,use-fpu is 1
  fpu-hard-extended    xlnx,use-fpu is 2
  vX.YY                model value of X.YY[.Z]

Signed-off-by: Mark Hatle <mark.hatle@xilinx.com>