meta-xilinx-bsp: Adjust soc to permit multiple CPU/TUNES
The various Xilinx FPGAs may have more then one CPU type including
cortexr5, microblaze and regular ARM Cortex CPUs.

Adding a new soc-tune-include.inc will allow the machine to choose a
default tune, and then the correct matching tune will be loaded.

In a perfect world this wouldn't be required, but doing it this way
permits us to target specific optimizations or CPUs in the soc.

Signed-off-by: Mark Hatle <mark.hatle@xilinx.com>