esw.bbclass: Only work with xilinx-standalone distro
Issue: CR-1058773

There are two versions of the plm recipe.  One that works with the
traditional Linux based build (basically includes a prebuilt version)
and one that is built from source using the standalone configuration.

There is currently no way to say a recipe is only supported when used
with a particular distribution, so we add code to the esw class that
makes it only usable with xilinx-standalone and xilinx-standalone-lto.

In the future, we may change the way this is done.

Signed-off-by: Mark Hatle <mark.hatle@xilinx.com>