Define COMPATIBLE_HOST to prevent mix of Linux and Baremetal recipes
The key is using COMPATIBLE_HOST to declare, for each recipe, which
OS they are compatible with.  This is a regex, so we need to
carefully specify the arch/vendor wild cardscomponents so that we match
only the OS parts.

In the distro configuration, default to the standard Linux HOST_SYS
setting.

In the various baremetal recipes (or esw class) set the specific
compatibility to allow baremetal recipes to build.

Signed-off-by: Mark Hatle <mark.hatle@xilinx.com>