microblaze machines: Set LINKER_HASH_STYLE defaults
For Linux based builds, it should be set to sysv.  This will result
in the system using a supported hash type.  (GNU hash is not supported
on Microblaze., so sysv will be used instead.)

For baremetal, set the value to "".  This will result in no value set,
and the system will use the compiler and linker default.

Signed-off-by: Mark Hatle <mark.hatle@xilinx.com>