machines: Remove default SERIAL_CONSOLES_CHECK
By default the machines should not check and remove declared consoles
that are not available on first boot.  It's up to the user to add this
to their build configuration.

The recommended user behavior is:

  SERIAL_CONSOLES_CHECK = ${SERIAL_CONSOLES}

but the side effect of this is that if the device configuration changes
after the first boot, the additional devices will not be available.

(Note, this may result in warning messages with getty unable to connect
to certain devices.)

Signed-off-by: Mark Hatle <mark.hatle@xilinx.com>