meta-xilinx-standalone gcc: Fix microblaze crtend.o
Issue: CR-1065455

Microblaze does not support initfini_array in a baremetal configuration.
This has previously been disabled in binutils, but it was missed in GCC and
related libraries.

This patch slightly restructures the baremetal configurations to move the
common configuration elements into a single configure set.  This matches the
behavior from the Yocto Project gcc-configure-common.inc file.

Signed-off-by: Mark Hatle <mark.hatle@xilinx.com>