meta-xilinx-bsp: binutils: Fix microblaze compilation
In some situations the microblaze patches will be enabled for the
binutils-native compilation.  Fix a compilation issue related to a
change in the way the md_pseudo_table structure definition has changed.

| ../../gas/config/tc-i386.c:1155:20: error: conflicting types for
‘md_pseudo_table’
| const pseudo_typeS md_pseudo_table[] =
|           ^~~~~~~~~~~~~~~
| In file included from ../../gas/as.h:565:0,
|         from ../../gas/config/tc-i386.c:28:
| ../../gas/tc.h:25:21: note: previous declaration of ‘md_pseudo_table’
was here| extern pseudo_typeS md_pseudo_table[];
|           ^~~~~~~~~~~~~~~

To support binutils-native on other host types, such as aarch64 a
similar patch is likely needed for those types as well.

Signed-off-by: Mark Hatle <mark.hatle@xilinx.com>