kernel-module-dp: Add support for Xilinx DP drivers
This recipe will pull in the Xilinx DP drivers and associated
components, for soft IP, to be built as an out-of-tree modules.

Following out-of-tree modules will be built

dp-rx (v4l2):
The DisplayPortPlay 1.4 Rx Subsystem is a feature-rich soft IP
incorporating all the necessary logic to properly interface with PHY
layers and provide DP decoding functionality. The subsystem is a
hierarchical IP that bundles a collection of DP RX-related IP
sub-cores and outputs them as a single IP.  The subsystem outputs
multi-pixel video to AXI4-Stream Protocol interface.

dp-tx (drm):
The DisplayPortPlay 1.4 Tx subsystem is a feature-rich soft IP
incorporating all the necessary logic to properly interface with
xilinx PHY layers and provide DP encoding functionality. The subsystem
is a hierarchical IP that bundles a collection of DP TX-related IP
sub-cores and outputs them as a single IP. The subsystem takes incoming
video and transfers them to an DP stream. The stream is then forwarded
to the video PHY layer.

video-phy (phy):
Xilinx Video Phy implements the physical layer for enabling the
plug-and-play connectivity with dp MAC transmit and receive
subsystems.  This driver is also used as the repository for common files
used by dp rx and tx drivers including hdcp1.4 and hdcp2.x software
stack.

Signed-off-by: Sandeep Gundlupet Raju <sandeep.gundlupet-raju@xilinx.com>
Signed-off-by: Sai Hari Chandana Kalluri <chandana.kalluri@xilinx.com>