Various: Packages using libGL may need to be declared SOC_VARIANT_ARCH
Issue: CR-1082769

libGL is either provided by mesa or libmali.  Each of the modified recipes
contains a direct library link to the provider of libGL.  Due to this we
need to specify precise dependencies and declare the package as the right
architecture to break a potential install conflict.

If we don't do this, the package will be generated as aarch64, but won't
be installable as the dependency on libmali can't be satisfied unless
the user has libmali, which is only available on zynqmpev and zyncmpeg.

git cherry-pick 6c0b7e5
Signed-off-by: Mark Hatle <mark.hatle@xilinx.com>