recipes-standalone: Move firmware to embeddedsw directory and simplify
Create a new embeddedsw.inc file,  This will contain all of the refereces
for SRCREV, and SRC_URI and other components that may be necessary.

Changing the filename or passing SRC_VER will switch the various source
version information.  As new releases happen, they can be added to the list
while incompatible ones can be removed.

This now allows us to focus on adjusting the recipes as necessary.

Signed-off-by: Mark Hatle <mark.hatle@kernel.crashing.org>