meta-xilinx-standalone: Prevent user error, when meta-microblaze is required
When the user chooses to build a microblaze target, the compiler can fail
to patch properly as a symptom.  Users may think it's a simple patch failure,
but really they forgot to include the meta-microblaze layer itself.

Declare the meta-microblaze layer as a recommended layer, and verify in the
gcc-source bbappend that it has been included, if a microblaze target is being
built.

Signed-off-by: Mark Hatle <mark.hatle@kernel.crashing.org>