xrt: Add run-time dependency on zocl
XRT has a run-time dependency on zocl for correct operation. Add zocl as
RDEPENDS.

Since XRT is machine independent, but zocl is tied to the kernel (because
it's a kernel module) and is thus kernel/machine specific, any time the
machine is changed, the hash for XRT would change.

To 'break' the dependency when it comes to hash calculations, add a
SIGGEN_EXCLUDE declaring this dependency as 'safe'.

Signed-off-by: Christian Kohn <christian.kohn@xilinx.com>
Signed-off-by: Sai Hari Chandana Kalluri <chandana.kalluri@xilinx.com>