embeddedsw: Rework plm/pmu/psm firmware and Linux packaging
In order to allow standalone (meta-xilinx-standalone), XSCT
(meta-xilinx-tools), and future items to work in the same way
the recipes have been restructured.

A *-firmware recipe will generate the firmware and stage it to do_deploy.
A *fw recipe will take the deployed version and package it for the Linux
side of things.  This allows the firmware generation to be easily extended
without requiring packaging knowledge.  Similarly packaging can be
extended for alternative boot/upgrade mechanisms as required.

In all cases, the MACHINE configuration will specify the default way
the components are to be built, along with the names of the item in
the deploy directory.

The PLM/PSM/PMU_IMAGE_NAME is the name for the generated firmware.

PLM/PSM/PMU_DEPLOY_DIR is the path to the constructed firmware.  This along
with the IMAGE_NAME above can be used to specify the location of an
externally generated set of firmware.

Addtionally the dependencies for building the plmfw/psmfw/pmufw can be
changed easily using PLM/PSM/PMU_DEPENDS and PLM/PSM/PMU_MCDEPENDS. The
former specifies dependencies in the same multiconfig, while the later
allows the component to require another multiconfig to have finihed.

The system has a referenced default, if multiconfig is enabled it will
automatically use it, otherwise it will try to use the recipe in the
main configuration.  (This will fail unless meta-xilinx-tools is available.)

Also two multiconfigs hve been implemented: versal-fw and zynqmp-pmufw
They can be enabled using BBMULITCONFIG += "zynqmp-pmufw" or versal-fw.

Signed-off-by: Mark Hatle <mark.hatle@xilinx.com>