u-boot-zynq-scr: update QSPI/NAND commands
In past using the if check in qspi/nand to check if user wants to use
fitimage/separate images default is fitimage, this leads to rebuild the boot.scr if user
wants to use separate images.
This patch will remove if check and use direct commands if proper image
found bootm will run and boots the kernel.

Signed-off-by: Raju Kumar Pothuraju <raju.kumar-pothuraju@xilinx.com>
Signed-off-by: Sai Hari Chandana Kalluri <chandana.kalluri@xilinx.com>