kernel-module-*: Do not set PACKAGE_ARCH to something less specific
Issue: CR-1098191

The kernel-module recipes set PACKAGE_ARCH = "${SOC_FAMILY_ARCH}" which is broken
because kernel modules depend on the kernel which is MACHINE specific. The
assignment claims that you could use a kernel module built for some other board
for any other zynqmp board, which is not true since their kernel versions and
configuration will differ.

Remove the assignment so that the kernel module becomes MACHINE_ARCH again.

This solves issues with kernel modules that refuse to install during the
do_rootfs stage, and various setscene tasks running over and again when
switching between MACHINEs.

Signed-off-by: Mike Looijmans <mike.looijmans@topic.nl>
Signed-off-by: Mark Hatle <mark.hatle@xilinx.com>