fsbl/fsbl-firmware: Split the packaging and building of the fsbl
Following the example of the psm, pmu and plm, we split the FSBL into both
a packaging component (fsbl) and the build of the firmware (fsbl-firmware).

This also adds an additional multilib, fsbl-fw, that is used to handle the
multiconfig version, as the Linux and baremetal builds are done differently.

Typical build approach is:

Add to local.conf:
 BBMULTICONFIG += "fsbl-fw"

then build using:
 MACHINE=zynqmp-generic bitbake fsbl

Note, while 'zynq' is implemented, it does not currently function.  To
build the fsbl for zynq, you must use the meta-xilinx-tools version.

Signed-off-by: Mark Hatle <mark.hatle@xilinx.com>