chromium-x11: Fix libmali package on zynqmp eg/ev
Issue: CR-1099895

libmali provides replacement items for standard mesa libraries.  Mali
also sets the soname to libMali.so, even when the requester is looking
for a library like libgbm.  This introduces an internal, unstated,
dependency on the provider of the mesa components.

Since this dependency change comes in from mali400, we need to
change the PACKAGE_ARCH in these few cases to match.

Signed-off-by: Mark Hatle <mark.hatle@xilinx.com>