cairo: Fix enabling egl and glesv2 on Mali400 platforms
Issue: CR-1099768

The original patch mistakenly tried to append for only zynqmp.  However,
it instead overwrote the original value.  Make this clear that we are
setting the value, not adding to an existing one.

Additionally, as the items are all mali400 specific, we ensure that the
dependency and PACKAGE_ARCH are set properly in this case.

Signed-off-by: Mark Hatle <mark.hatle@xilinx.com>