weston: The package uses different sources when build on zynqmp
Issue: CR-1099768

Combine two different weston bbappends to avoid confusion and conflicts
between the two.

Adjust the parameters of the original bbappend to use the mali400 override,
as well as define a new zynqmp (non-mali) PACKAGE_ARCH due to the SRC_URI
patch that is also zynqmp specific.

The end result will be:
  not-zynqmp:  TUNE_PKGARCH
  zynqmp w/o mali: SOC_FAMILY_ARCH
  zynqmp w/ mali: SOC_VARIANT_ARCH

Signed-off-by: Mark Hatle <mark.hatle@xilinx.com>