Change the order of the MACHINEOVERRIDES to ensure consistent behavior
Issue: CR-1100158

Previously the final order of the MACHINEOVERRIDES was effectively:

 zynqmp_graphics:board_variant:board:soc_variant:soc:aarch64:armv8a:cortexa72-cortexa53:machine

The correct order should be least important to most important:

 aarch64:armv8a:cortexa72-cortexa53:zynqmp_graphics:soc:soc_variant:machine:board:board_variant

This change accomplishes that by adjusting the order of the .= / =. of the
MACHINEOVERRIDES values.  This changes the order of a few standard requires,
adjusts how the zynqmp_graphics is added, and changes the order of the
soc/soc_variant settings.

Signed-off-by: Mark Hatle <mark.hatle@xilinx.com>