linux-firmware: Move from bbappend to a custom firmware recipe
Ultra96 WiFi/BT chip requires custom firmware.  The main wifi firmware is
provided by the linux-firmware recipe, but the Bluetooth firmware must be
sourced from another location.

We only package the specific firmware that we require for this, and do
it in a way that is similar to the linux-firmware recipe.

Signed-off-by: Mark Hatle <mark.hatle@xilinx.com>