meta-xilinx-standalone-experimental: Move from device-tree-lops to device-tree
The device trees are now generated by the user as part of the configuration
steps.  As such, we need to simply copy/process these device trees as part
of the system build so adjust the various usages to the new variable
CONFIG_DTFILE, which is a full path reference to the device tree for a
particular configuration.

Signed-off-by: Mark Hatle <mark.hatle@xilinx.com>