meta-xilinx-setup: Update dt-processor.sh
The Linux build must ONLY refer to the Linux specific device tree or boot
and QEMU failures can occur.  The issue was resolved by ensuring that the
system always considered the Linux device tree as the "system" device tree
when building for the Linux OS.  Additionally, as multiple device trees
may be present, re-order the IMAGE_BOOT_FILES to ensure the correct
device tree is listed first.  This ensures that both QEMU and the u-boot
configuration are referring to the correct device tree.

Signed-off-by: Mark Hatle <mark.hatle@xilinx.com>