meta-xilinx-setup: Add BASE_TMPDIR
A new variable BASE_TMPDIR was defined to make it easier for a user to choose
an alternative directory for the various TMPDIR files.  Since there are
references from one TMPDIR to another, we need a simply way to adjust them
all as a unit.

A developer should NOT change the actual TMPDIR specifications, unless they
also change all of the other generated configurations!

Signed-off-by: Mark Hatle <mark.hatle@xilinx.com>