microblaze m4/diffutils/grep: Add stack direction info to m4 macros
These tools require knowledge of the stack direction, since we are cross
compiling they are unable to detect the stack growth direction so it must
be defined in the m4 macro.

It turns out the direction was already set, but defined as
'microblaze' instead of 'microblaze*' which caused problems on little
endian.

Signed-off-by: Mark Hatle <mark.hatle@xilinx.com>