Adjust generic BSPs to support the inheritance model
Each inherited BSP must set a MACHINEOVERRIDES and PACKAGE_EXTRA_ARCHS to
ensure that various overrides can be set along the way.  The general format
for this is (in the caller):

require conf/machine/<inherited machine>.conf

The callee is responsible for adding to the MACHINEOVERRIDES and
PACKAGE_EXTRA_ARCHS.  It does this by setting the following at the beginning
and end of the file, respectively.

MACHINEOVERRIDES =. "${@['', '<machine>:']['<machine>' != '${MACHINE}']}"
...
...
PACKAGE_EXTRA_ARCHS_append = "${@['', ' <machine_arch>']['<machine>' != '${MACHINE}']}"

where <machine> is the <machine>.conf file the line appears in;
<machine_arch> is the <machine>.replace('-', '_') (matching MACHINE_ARCH) if
the machine file was used by itself.

The MACHINEOVERRIDES uses =. or .= everywhere in the Yocto Project.  Doing it
via =. in the load order ensures in order evaluation.

PACKAGE_EXTRA_ARCHS is set via 'append', but needs to be the last operation
to ensure that we are always added AFTER anything we include along the way.
Effectively this is a late binding evaluation.

Signed-off-by: Mark Hatle <mark.hatle@xilinx.com>