device-tree.bbappend: defining the memory and serial console param
This patch will define the device-tree required memory and uart params.
Due to recent board/board_variant and machine changes in yocto caused
this issue.To work with these new changes we need this.
In plnxtool.conf we will define the default values but if we set board
value in that case due to the changes in new microblaze-generic.conf file
the plnxtool.conf changes will not work.

Signed-off-by: Varalaxmi Bingi <varalaxmi.bingi@xilinx.com>
Signed-off-by: Mark Hatle <mark.hatle@xilinx.com>