device-tree: Merge experimental and standard flows
We need a method for regular building to specify a device tree.  Since the
experimental flow already contained this, merge it in and connect to
the existing SYSTEM_DTFILE as a fall back.

Also enable addition 'early package failure' messages to alert the user as
soon as possible they need to pass in additional information.

Signed-off-by: Mark Hatle <mark.hatle@xilinx.com>