embeddedsw: Rework the embeddedsw copy firmware recipes
fsbl, plmfw, pufw, and psmfw are reworked to provide earlier error messages,
and better handle multiconfig builds.

Also adjust MACHINE specific changes to use .* for compatibility based on the
SOC_FAMILY override.  The system doesn't always set the SOC_FAMILY as
COMPATIBLE_MACHINE, so this avoids an error of incompatible machine.

Align all of the expected firmware filenames to be <firmware>-${MACHINE}.
The user can override as necessary.

The license moves to CLOSED as we don't know the license of the software being
provided by the user or other dependencies.  It'll be up to the user to
reconcile it.

Signed-off-by: Mark Hatle <mark.hatle@xilinx.com>