xilinx-bootbin: Sync versal and ensure items can be defined in local.conf
Use ??= to define the default values.  This ensures a ?= (or =) can be
used in a .bbappend, or a user can define a default value in the local.conf
or machine.conf file.

Also add a default config for versal that matches zynqmp and zynq.

Signed-off-by: Mark Hatle <mark.hatle@xilinx.com>
(cherry picked from commit a32c040885f065c94be257a7dd92daefe76b3435)
Signed-off-by: Mark Hatle <mark.hatle@xilinx.com>