weston-init: Pass --continue-without-input when launching weston
This ensures that we do not need keyboard/mouse or other input system
Weston's default behavior is to not open if input devices are not
found, but we may not always have input devices connected for systems
running weston in field.

Signed-off-by: Khem Raj <raj.khem@gmail.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>

This commit is based on 762a20b493cc219a46d9ac188fe4895a111ee7b4 from
meta-openembedded-core. It was later reverted in master because it breaks
the fbdev backend. At Xilinx, we are using the drm backend so it is safe
to add this.

Signed-off-by: Christian Kohn <christian.kohn@xilinx.com>
Signed-off-by: Mark Hatle <mark.hatle@xilinx.com>
(cherry picked from commit bb7439368c6a1fc17ccb32731471d9e6c5661ba6)
Signed-off-by: Mark Hatle <mark.hatle@xilinx.com>