meta-xilinx-standalone-experimental: fsbl-firmware: Fix issue with fetch into wrong directory
The file:// items were being placed in the same path structure in WORKDIR
as they were in the original path passed into the recipe.  Instead we want
the files to always just show up in the WORKDIR.

Move to using a simple file://psu_init.[ch] entry, but use FILESEXTRAPATHS
to handle the access to the correct path.  Upside is that everything works
as expected, downside, if the path to the psu_init files also contain other
items refered to by SRC_URI, it could pick up that version instead.

With the current implementation, this is unlikely to be a problem.

The code also moves the warning about using the default psu_init files to
the do_compile.  Unfortunately doing it in the anonymous python causes
the warning to be generated for all multiconfigs, even ones that will never
build the fsbl files.  There is no simply way to disable this warning in
those cases.

Signed-off-by: Mark Hatle <mhatle@xilinx.com>
(cherry picked from commit 49912c9a380bd1b472abe925e83b785b16e76dfb)
Signed-off-by: Mark Hatle <mhatle@xilinx.com>